Refactor runner classes
