--- masterFile.py
@@ -1,7 +1,0 @@
-#the main file for the solution
-
-if __name__ == "__main__":
-    from Solutions2020.y2020d16 import y2020d16
-
-    print(y2020d16())
-
