--- runSolution.bat
@@ -1,2 +1,0 @@
-python masterFile.py
-cmd -k
